requirements.txt
ragas==0.1.9
datasets==2.14.6
langchain==0.1.0
langchain-core==0.1.52
langchain-google-genai==2.0.8
langchain-huggingface==0.1.2
google-generativeai==0.8.3
sentence-transformers==3.3.1
pandas==2.1.3
numpy==1.24.3
